Gunnar: compatibility issue
git-svn-id: https://www.nick-ux.org/diagham/svn/trunk@1582 b9410c57-6ef9-0310-98bc-a73ba792996a

+++ Spin/src/Hamiltonian/SUNSpinOnLatticeQuadraticHamiltonian.cc
@@ -134,7 +134,7 @@ void SUNSpinOnLatticeQuadraticHamiltonian::EvaluateInteractionTerms()
 	  ++Pos;
 	}
     }
-  if (fabs(this->CyclicPrefactor>0.0))
+  if (fabs((double)this->CyclicPrefactor)>0.0)
     {
       this->NbrCyclicPermutations = 2*this->Lattice->GetNbrPlaquettes();
       if (this->NbrCyclicPermutations==0)
